feat: Add new rolling stats and break out more functions
